backend/requirements.txt
# Pin rationale: original spec had langgraph==0.2.0 + langchain-core==0.3.0 —
# unsolvable (langgraph 0.2.0 requires core<0.3). First bump (langgraph==0.2.60
# + core==0.3.15) also failed because langgraph 0.2.60 excludes core 0.3.0–0.3.22.
# Final combo uses ranges for the langchain trio so pip resolves the latest
# compatible set; other packages stay exactly pinned. See MEMORY.md for log.
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai>=1.40.0,<2
langgraph>=0.2.60,<0.3
langchain-openai>=0.2,<0.3
langchain-core>=0.3.23,<0.4
supabase==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
httpx==0.27.0
gdown>=5.1.0,<6
PyJWT>=2.8,<3
cryptography>=42,<46
pytest==8.3.0
pytest-asyncio==0.24.0
